refactor: :zap: add minors changes
fix typos and update minor styles, all test passed
